peace_map/api/requirements.txt
# Peace Map API Requirements

# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Geospatial
geopy==2.4.1
folium==0.15.0

# Machine Learning
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Background tasks
celery==5.3.4
redis==5.0.1

# NLP
nltk==3.8.1
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
